refactor: удалить InMemory классы и тесты, переместить ControllerExceptionHandler в exceptions
